po-automation/requirements.txt
anthropic
streamlit
pandas
xlrd==1.2.0
openpyxl
pdfplumber
imapclient
email
python-dotenv
